code-samples/calendar/python/create_calendar_event/pyproject.toml
[tool.poetry]
name = "create-calendar-event"
version = "0.1.0"
description = "Create a calendar event using Nylas Python SDK"
authors = ["wiseaidev <oss@wiseai.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nylas = "^5.14.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
pytest = "^7.4.2"
isort = "^5.12.0"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
create_calendar_event = "create_calendar_event.__main__:main"

[virtualenvs]
in-project = true
create = true